refactor(main): split __main__ into lyra/bootstrap package
__main__.py reduced from 1406 → 187 lines. Extracted sub-domains:
- bootstrap/config.py      — config loaders (_load_raw_config, etc.)
- bootstrap/agent_factory.py — agent creation & resolution
- bootstrap/health.py      — FastAPI health + config endpoints
- bootstrap/multibot.py    — multi-bot startup wiring
- bootstrap/legacy.py      — legacy single-bot startup wiring
- bootstrap/utils.py       — shared task failure callback

Also fixed double-close bug: stores were closed in both normal
shutdown path and finally block; unified to finally-only.

Updated all tests to import from new module locations.
