+++ src/lyra/__main__.py
@@ -7,1282 +7,35 @@ from __future__ import annotations
 
 import argparse
 import asyncio
-import dataclasses
-import hmac
 import logging
 import os
-import re
-import signal
 import sys
-import time
-import tomllib
 from datetime import datetime, timezone
 from logging.handlers import RotatingFileHandler
 from pathlib import Path
 
-import uvicorn
 from dotenv import load_dotenv
-from fastapi import FastAPI, Header, HTTPException
 
-from lyra.adapters.discord import DiscordAdapter
-from lyra.adapters.telegram import TelegramAdapter
-from lyra.agents.simple_agent import SimpleAgent
+# Re-export adapters so test monkeypatching on main_mod still works.
+# The bootstrap modules import these directly; tests that patch main_mod.*
+# must also patch the bootstrap module namespace (see tests/test_main.py).
+from lyra.adapters.discord import DiscordAdapter  # noqa: F401
+from lyra.adapters.telegram import TelegramAdapter  # noqa: F401
+from lyra.bootstrap.config import (
+    _load_circuit_config,
+    _load_raw_config,
+)
+from lyra.bootstrap.legacy import _bootstrap_legacy
+from lyra.bootstrap.multibot import _bootstrap_multibot
 from lyra.config import (
-    DiscordBotConfig,
     DiscordMultiConfig,
-    TelegramBotConfig,
     TelegramMultiConfig,
     load_multibot_config,
 )
 from lyra.core.admin import set_admin_user_ids
-from lyra.core.agent import (
-    Agent,
-    AgentBase,
-    AgentSTTConfig,
-    AgentTTSConfig,
-    SmartRoutingConfig,
-    agent_row_to_config,
-    load_agent_config,
-)
-from lyra.core.agent_store import AgentStore
-from lyra.core.auth import AuthMiddleware
-from lyra.core.auth_store import AuthStore
-from lyra.core.circuit_breaker import CircuitBreaker, CircuitRegistry
-from lyra.core.cli_pool import CliPool
-from lyra.core.credential_store import CredentialStore, LyraKeyring
-from lyra.core.hub import Hub, RoutingKey
-from lyra.core.message import Platform
-from lyra.core.messages import MessageManager
-from lyra.core.outbound_dispatcher import OutboundDispatcher
-from lyra.core.pairing import PairingConfig, PairingManager, set_pairing_manager
-from lyra.core.prefs_store import PrefsStore
-from lyra.core.thread_store import ThreadStore
-from lyra.core.turn_store import TurnStore
 from lyra.errors import KeyringError, MissingCredentialsError
-from lyra.llm.base import LlmProvider
-from lyra.llm.registry import ProviderRegistry
-from lyra.llm.smart_routing import SmartRoutingDecorator
-from lyra.stt import STTConfig, STTService, load_stt_config
-from lyra.tts import TTSConfig, TTSService, load_tts_config
 
 log = logging.getLogger(__name__)
-
-
-def _log_task_failure(task: asyncio.Task) -> None:  # type: ignore[type-arg]
-    """Done callback — logs adapter task failures (e.g. Discord LoginFailure)."""
-    if task.cancelled():
-        return
-    try:
-        exc = task.exception()
-    except asyncio.CancelledError:
-        return
-    if exc is not None:
-        log.error(
-            "Adapter task '%s' exited with error: %s",
-            task.get_name(),
-            exc,
-            exc_info=exc,
-        )
-
-
-_CB_DEFAULTS: dict[str, int] = {"failure_threshold": 5, "recovery_timeout": 60}
-_CB_SERVICES = ("anthropic", "telegram", "discord", "hub")
-_ADMIN_ID_PATTERN = re.compile(r"^(tg|dc):user:\d+$")
-
-
-def _load_raw_config(config_path: str | None = None) -> dict:
-    """Open and parse config.toml once; return the raw dict.
-
-    Resolution order: $LYRA_CONFIG env var → 'config.toml' in cwd → empty dict.
-    """
-    raw: dict = {}
-    path = config_path or os.environ.get("LYRA_CONFIG", "config.toml")
-    try:
-        with open(path, "rb") as f:
-            raw = tomllib.load(f)
-        log.info("Loaded config from %s", path)
-    except FileNotFoundError:
-        log.info("No config file at %s — using defaults", path)
-    return raw
-
-
-def _build_agent_overrides(raw: dict, name: str) -> dict:
-    """Build instance-level overrides for an agent from config.toml.
-
-    Merges [defaults] with [agents.<name>], with agent-specific values winning.
-    Provides machine-specific fallbacks for cwd, persona, and workspaces that
-    are not versioned in agents/*.toml.
-    """
-    defaults: dict = raw.get("defaults", {})
-    agent_specific: dict = raw.get("agents", {}).get(name, {})
-    merged = {**defaults, **agent_specific}
-    # workspaces need a deep merge: defaults provide base, agent-specific keys win
-    ws_defaults = defaults.get("workspaces", {})
-    ws_agent = agent_specific.get("workspaces", {})
-    if ws_defaults or ws_agent:
-        merged["workspaces"] = {**ws_defaults, **ws_agent}
-    return merged
-
-
-def _load_circuit_config(
-    raw: dict,
-) -> tuple[CircuitRegistry, set[str]]:
-    """Load [circuit_breaker.*] and [admin] sections from raw config dict.
-
-    Returns (CircuitRegistry with 4 named CBs, set of admin user_ids).
-    Missing sections → all defaults.
-    """
-    cb_section = raw.get("circuit_breaker", {})
-    registry = CircuitRegistry()
-    for name in _CB_SERVICES:
-        cfg: dict[str, int] = {**_CB_DEFAULTS, **cb_section.get(name, {})}
-        registry.register(
-            CircuitBreaker(
-                name=name,
-                failure_threshold=cfg["failure_threshold"],
-                recovery_timeout=cfg["recovery_timeout"],
-            )
-        )
-
-    admin_ids: set[str] = set(raw.get("admin", {}).get("user_ids", []))
-    for aid in admin_ids:
-        if not _ADMIN_ID_PATTERN.match(aid):
-            log.warning(
-                "Admin ID %r does not match expected format "
-                "(tg|dc):user:<digits> — verify config.toml [admin].user_ids",
-                aid,
-            )
-    return registry, admin_ids
-
-
-def _load_pairing_config(raw: dict) -> PairingConfig:
-    """Load [pairing] section from raw config dict. Missing section → all defaults."""
-    pairing_section: dict = raw.get("pairing", {})
-    return PairingConfig.from_dict(pairing_section)
-
-
-def _load_messages(language: str = "en") -> MessageManager:
-    """Load MessageManager.
-
-    Resolution order:
-      LYRA_MESSAGES_CONFIG env var → messages.toml in cwd → bundled config.
-    """
-    bundled = Path(__file__).resolve().parent / "config" / "messages.toml"
-    path_str = os.environ.get("LYRA_MESSAGES_CONFIG") or (
-        "messages.toml" if Path("messages.toml").exists() else str(bundled)
-    )
-    if not path_str.endswith(".toml"):
-        log.warning(
-            "LYRA_MESSAGES_CONFIG %r does not end with .toml — ignoring, using bundled",
-            path_str,
-        )
-        path_str = str(bundled)
-    log.info("Loaded messages from %s", path_str)
-    return MessageManager(path_str, language=language)
-
-
-def _build_provider_registry(
-    circuit_registry: CircuitRegistry,
-    cli_pool: CliPool | None,
-    smart_routing_config: SmartRoutingConfig | None = None,
-) -> tuple[ProviderRegistry, SmartRoutingDecorator | None]:
-    """Build and return a ProviderRegistry with all configured drivers.
-
-    Returns (registry, smart_routing_decorator_or_None) so the routing
-    decorator's history is accessible to the /routing admin command.
-    """
-    from lyra.llm.decorators import CircuitBreakerDecorator, RetryDecorator
-
-    registry = ProviderRegistry()
-    routing_decorator: SmartRoutingDecorator | None = None
-
-    if cli_pool is not None:
-        from lyra.llm.drivers.cli import ClaudeCliDriver
-
-        registry.register("claude-cli", ClaudeCliDriver(cli_pool))
-        log.info("ProviderRegistry: registered claude-cli driver")
-
-    api_key = os.environ.get("ANTHROPIC_API_KEY")
-    if api_key:
-        from lyra.llm.drivers.sdk import AnthropicSdkDriver
-
-        sdk_driver = AnthropicSdkDriver(api_key)
-        retry = RetryDecorator(sdk_driver, max_retries=3, backoff_base=1.0)
-
-        # Smart routing wraps retry (CB → SmartRouting → Retry → Driver)
-        inner: LlmProvider = retry
-        if smart_routing_config is not None:
-            from lyra.llm.smart_routing import SmartRoutingDecorator as SRD
-
-            routing_decorator = SRD(retry, smart_routing_config)
-            inner = routing_decorator
-
-        anthropic_cb = circuit_registry.get("anthropic")
-        if anthropic_cb is not None:
-            provider: LlmProvider = CircuitBreakerDecorator(inner, anthropic_cb)
-        else:
-            provider = inner
-        registry.register("anthropic-sdk", provider)
-        log.info("ProviderRegistry: registered anthropic-sdk driver (decorated)")
-
-    return registry, routing_decorator
-
-
-def _create_agent(  # noqa: PLR0913 — factory with optional overrides for each agent dependency
-    config: Agent,
-    cli_pool: CliPool | None,
-    circuit_registry: CircuitRegistry | None = None,
-    admin_user_ids: set[str] | None = None,
-    msg_manager: MessageManager | None = None,
-    stt: STTService | None = None,
-    tts: TTSService | None = None,
-    provider_registry: ProviderRegistry | None = None,
-    smart_routing_decorator: SmartRoutingDecorator | None = None,
-) -> AgentBase:
-    """Select agent implementation based on backend config."""
-    backend = config.model_config.backend
-    if backend == "anthropic-sdk":
-        from lyra.agents.anthropic_agent import AnthropicAgent
-
-        if provider_registry is not None:
-            provider = provider_registry.get("anthropic-sdk")
-        else:
-            from lyra.llm.drivers.sdk import AnthropicSdkDriver
-
-            api_key = os.environ.get("ANTHROPIC_API_KEY")
-            if not api_key:
-                raise SystemExit("Missing required env var: ANTHROPIC_API_KEY")
-            provider = AnthropicSdkDriver(api_key)
-        return AnthropicAgent(
-            config,
-            provider,
-            circuit_registry=circuit_registry,
-            admin_user_ids=admin_user_ids,
-            msg_manager=msg_manager,
-            stt=stt,
-            tts=tts,
-            smart_routing_decorator=smart_routing_decorator,
-        )
-    if backend in ("claude-cli", "ollama"):
-        if cli_pool is None:
-            raise RuntimeError(f"CliPool required for {backend} backend")
-        if provider_registry is not None:
-            provider = provider_registry.get("claude-cli")
-        else:
-            from lyra.llm.drivers.cli import ClaudeCliDriver
-
-            provider = ClaudeCliDriver(cli_pool)
-        return SimpleAgent(
-            config,
-            provider,
-            circuit_registry=circuit_registry,
-            admin_user_ids=admin_user_ids,
-            msg_manager=msg_manager,
-            stt=stt,
-            tts=tts,
-        )
-    raise ValueError(f"Unknown backend: {backend}")
-
-
-def create_health_app(hub: Hub) -> FastAPI:
-    """Create a root FastAPI app with /health endpoint for hub monitoring.
-
-    This is the top-level HTTP app — adapter sub-apps can be mounted on it.
-    The /health endpoint exposes hub-level health without requiring adapter auth.
-    """
-    app = FastAPI(title="Lyra Hub")
-
-    @app.get("/health")
-    async def health(authorization: str = Header(default="")) -> dict:
-        health_secret = os.environ.get("LYRA_HEALTH_SECRET", "")
-        expected = f"Bearer {health_secret}"
-        authenticated = bool(health_secret) and hmac.compare_digest(
-            authorization, expected
-        )
-
-        # Security: wrong/missing token intentionally returns the minimal
-        # response rather than 401, to avoid revealing whether a secret is
-        # configured.  This differs from /config which returns 401.
-        if not authenticated:
-            return {"ok": True}
-
-        uptime_s = time.monotonic() - hub._start_time
-
-        last_message_age_s: float | None = None
-        if hub._last_processed_at is not None:
-            last_message_age_s = time.monotonic() - hub._last_processed_at
-
-        circuits: dict[str, dict[str, object]] = {}
-        if hub.circuit_registry is not None:
-            all_status = hub.circuit_registry.get_all_status()
-            circuits = {
-                name: {
-                    "state": s.state.value,
-                    "retry_after": s.retry_after,
-                }
-                for name, s in all_status.items()
-            }
-
-        inbound: dict[str, int] = {
-            p.value: hub.inbound_bus.qsize(p)
-            for p in hub.inbound_bus.registered_platforms()
-        }
-        outbound: dict[str, int] = {
-            platform.value: dispatcher.qsize()
-            for (platform, _bot_id), dispatcher in hub.outbound_dispatchers.items()
-        }
-
-        return {
-            "ok": True,
-            "queue_size": hub.inbound_bus.staging_qsize(),
-            "queues": {"inbound": inbound, "outbound": outbound},
-            "last_message_age_s": last_message_age_s,
-            "uptime_s": round(uptime_s, 1),
-            "circuits": circuits,
-        }
-
-    @app.get("/config")
-    async def config_endpoint(
-        authorization: str = Header(default=""),
-        agent: str = "lyra_default",
-    ) -> dict:
-        config_secret = os.environ.get("LYRA_CONFIG_SECRET", "")
-        if not config_secret or not hmac.compare_digest(
-            authorization, f"Bearer {config_secret}"
-        ):
-            raise HTTPException(status_code=401, detail="unauthorized")
-        from lyra.agents.anthropic_agent import AnthropicAgent
-
-        agent_obj = hub.agent_registry.get(agent)
-        if not isinstance(agent_obj, AnthropicAgent):
-            raise HTTPException(
-                status_code=404,
-                detail="runtime config not available for this agent backend",
-            )
-        rc = agent_obj.runtime_config
-        return {
-            "style": rc.style,
-            "language": rc.language,
-            "temperature": rc.temperature,
-            "model": rc.model,
-            "max_steps": rc.max_steps,
-            "extra_instructions": rc.extra_instructions,
-            "effective_model": rc.model or agent_obj.config.model_config.model,
-            "effective_max_steps": (
-                rc.max_steps or agent_obj.config.model_config.max_turns
-            ),
-        }
-
-    return app
-
-
-def apply_agent_tts_overlay(
-    agent_tts: AgentTTSConfig | None,
-    tts_cfg: TTSConfig,
-) -> TTSConfig:
-    """Overlay non-None fields from AgentTTSConfig onto TTSConfig.
-
-    None fields in agent_tts are skipped — voicecli global defaults remain in effect.
-    Returns an updated TTSConfig (via dataclasses.replace).
-    """
-    if agent_tts is None:
-        return tts_cfg
-    a = agent_tts
-    if a.engine is not None:
-        tts_cfg = dataclasses.replace(tts_cfg, engine=a.engine)
-    if a.voice is not None:
-        tts_cfg = dataclasses.replace(tts_cfg, voice=a.voice)
-    if a.language is not None:
-        tts_cfg = dataclasses.replace(tts_cfg, language=a.language)
-    return tts_cfg
-
-
-def apply_agent_stt_overlay(
-    agent_stt: AgentSTTConfig | None,
-    stt_cfg: STTConfig,
-) -> STTConfig:
-    """Overlay non-None fields from AgentSTTConfig onto STTConfig.
-
-    None fields in agent_stt are skipped — voicecli global defaults remain in effect.
-    Returns an updated STTConfig (via dataclasses.replace).
-    """
-    if agent_stt is None:
-        return stt_cfg
-    a = agent_stt
-    if a.language_detection_threshold is not None:
-        stt_cfg = dataclasses.replace(
-            stt_cfg, language_detection_threshold=a.language_detection_threshold
-        )
-    if a.language_detection_segments is not None:
-        stt_cfg = dataclasses.replace(
-            stt_cfg, language_detection_segments=a.language_detection_segments
-        )
-    if a.language_fallback is not None:
-        stt_cfg = dataclasses.replace(stt_cfg, language_fallback=a.language_fallback)
-    return stt_cfg
-
-
-def _resolve_agents(  # noqa: PLR0913
-    agent_configs: dict[str, Agent],
-    cli_pool: CliPool | None,
-    circuit_registry: CircuitRegistry,
-    admin_user_ids: set[str],
-    msg_manager: MessageManager,
-    stt_service: STTService | None,
-    tts_service: TTSService | None = None,
-) -> dict[str, AgentBase]:
-    """Create all uniquely named agents referenced by bot configs.
-
-    Builds a per-agent ProviderRegistry (and SmartRoutingDecorator when configured)
-    so that each agent's routing settings are respected independently.
-    Accepts pre-loaded agent configs to avoid duplicate I/O.
-    Returns a dict mapping agent_name → AgentBase instance.
-    """
-    agents: dict[str, AgentBase] = {}
-    for name, agent_config in sorted(agent_configs.items()):  # deterministic log order
-        log.info(
-            "Agent loaded: name=%s model=%s backend=%s",
-            agent_config.name,
-            agent_config.model_config.model,
-            agent_config.model_config.backend,
-        )
-        # Build per-agent provider registry so each agent's smart_routing config
-        # is applied independently (avoids silent cross-agent routing mis-dispatch).
-        sr_config = agent_config.smart_routing
-        if (
-            sr_config is not None
-            and sr_config.enabled
-            and agent_config.model_config.backend != "anthropic-sdk"
-        ):
-            log.warning(
-                "agent %r: smart_routing.enabled=true but backend=%r — "
-                "smart routing only applies to anthropic-sdk",
-                name,
-                agent_config.model_config.backend,
-            )
-        per_agent_registry, per_agent_routing = _build_provider_registry(
-            circuit_registry, cli_pool, smart_routing_config=sr_config
-        )
-        agent = _create_agent(
-            agent_config,
-            cli_pool,
-            circuit_registry=circuit_registry,
-            admin_user_ids=admin_user_ids,
-            msg_manager=msg_manager,
-            stt=stt_service,
-            tts=tts_service,
-            provider_registry=per_agent_registry,
-            smart_routing_decorator=per_agent_routing,
-        )
-        agents[name] = agent
-    return agents
-
-
-async def _resolve_bot_agent_map(
-    agent_store: AgentStore,
-    tg_bots: "list[TelegramBotConfig]",
-    dc_bots: "list[DiscordBotConfig]",
-) -> "dict[tuple[str, str], str]":
-    """Resolve (platform, bot_id) → agent_name for all configured bots.
-
-    Resolution order per bot:
-      1. bot_agent_map DB row (agent_store.get_bot_agent)
-      2. bot_cfg.agent from TOML — auto-seeds bot_agent_map in DB
-      3. Neither → log error, skip (bot not included in result)
-
-    Bots whose resolved agent name is not found in the agents table are also
-    logged and skipped (adapter cannot be wired without a valid agent row).
-
-    Returns dict mapping (platform, bot_id) → agent_name for bots that can
-    be safely wired.
-    """
-    result: dict[tuple[str, str], str] = {}
-
-    all_bots: list[tuple[str, str, str | None]] = []
-    for bot_cfg in tg_bots:
-        toml_agent = getattr(bot_cfg, "agent", None)
-        all_bots.append(("telegram", bot_cfg.bot_id, toml_agent))
-    for bot_cfg in dc_bots:
-        toml_agent = getattr(bot_cfg, "agent", None)
-        all_bots.append(("discord", bot_cfg.bot_id, toml_agent))
-
-    for platform, bot_id, toml_agent in all_bots:
-        # 1. Check DB cache (bot_agent_map table)
-        db_agent_name = agent_store.get_bot_agent(platform, bot_id)
-
-        if db_agent_name is not None:
-            # DB row found — validate the referenced agent exists in agents table.
-            # A stale bot_agent_map row (agent deleted) is a configuration error.
-            if agent_store.get(db_agent_name) is None:
-                log.error(
-                    "bot_agent_map: agent %r for (%r, %r) not found in"
-                    " agents table — skipping adapter",
-                    db_agent_name,
-                    platform,
-                    bot_id,
-                )
-                continue
-            result[(platform, bot_id)] = db_agent_name
-        elif toml_agent:
-            # Check agent exists in DB before seeding the mapping
-            if agent_store.get(toml_agent) is None:
-                log.error(
-                    "bot_agent_map: TOML agent %r for (%r, %r) not found in agents DB "
-                    "— skipping adapter (run 'lyra agent init' to import TOMLs)",
-                    toml_agent,
-                    platform,
-                    bot_id,
-                )
-                continue
-            # 2. No DB row — fall back to TOML bot_cfg.agent, seed bot_agent_map.
-            log.info(
-                "bot_agent_map: no DB row for (%r, %r) — seeding from TOML agent=%r",
-                platform,
-                bot_id,
-                toml_agent,
-            )
-            try:
-                await agent_store.set_bot_agent(platform, bot_id, toml_agent)
-            except Exception as exc:  # noqa: BLE001
-                log.warning(
-                    "bot_agent_map: failed to seed (%r, %r) → %r: %s",
-                    platform,
-                    bot_id,
-                    toml_agent,
-                    exc,
-                )
-            result[(platform, bot_id)] = toml_agent
-        else:
-            # 3. No DB row, no TOML agent — skip
-            log.error(
-                "bot_agent_map: no DB row and no TOML agent for"
-                " (%r, %r) — skipping adapter",
-                platform,
-                bot_id,
-            )
-            continue
-
-    return result
-
-
-async def _bootstrap_multibot(  # noqa: C901, PLR0915 — startup wiring: each adapter/service requires sequential conditional setup
-    raw_config: dict,
-    circuit_registry: CircuitRegistry,
-    admin_user_ids: set[str],
-    tg_multi_cfg: TelegramMultiConfig,
-    dc_multi_cfg: DiscordMultiConfig,
-    *,
-    _stop: asyncio.Event | None = None,
-) -> None:
-    """Wire hub + adapters for the multi-bot configuration schema and run until stop.
-
-    Handles [[telegram.bots]] / [[discord.bots]] arrays from config.toml.
-    Each bot is wired individually with its own AuthMiddleware, OutboundDispatcher,
-    and adapter.
-    """
-    # ------------------------------------------------------------------
-    # Multi-bot path: validate auth per bot, collect unique agent names.
-    # ------------------------------------------------------------------
-    vault_dir_mb = Path(os.environ.get("LYRA_VAULT_DIR", str(Path.home() / ".lyra")))
-    vault_dir_mb.mkdir(parents=True, exist_ok=True)
-    auth_store: AuthStore | None = None
-    cred_store: CredentialStore | None = None
-    agent_store: AgentStore | None = None
-    turn_store_mb: TurnStore | None = None
-    thread_store_mb: ThreadStore | None = None
-    try:
-        auth_store = AuthStore(db_path=vault_dir_mb / "auth.db")
-        await auth_store.connect()
-        keyring_mb = LyraKeyring.load_or_create(vault_dir_mb / "keyring.key")
-        cred_store = CredentialStore(
-            db_path=vault_dir_mb / "auth.db",
-            keyring=keyring_mb,
-        )
-        await cred_store.connect()
-        agent_store = AgentStore(db_path=vault_dir_mb / "auth.db")
-        await agent_store.connect()
-        turn_store_mb = TurnStore(db_path=vault_dir_mb / "turns.db")
-        await turn_store_mb.connect()
-        thread_store_mb = ThreadStore(db_path=vault_dir_mb / "auth.db")
-        await thread_store_mb.connect()
-        # Seed per-bot owner/trusted users as permanent grants
-        auth_block: dict = raw_config.get("auth", {})
-        for entry in auth_block.get("telegram_bots", []):
-            synthetic = {"auth": {"telegram": entry}}
-            await auth_store.seed_from_config(synthetic, "telegram")
-        for entry in auth_block.get("discord_bots", []):
-            synthetic = {"auth": {"discord": entry}}
-            await auth_store.seed_from_config(synthetic, "discord")
-
-        tg_bot_auths: list[tuple[TelegramBotConfig, AuthMiddleware]] = []
-        dc_bot_auths: list[tuple[DiscordBotConfig, AuthMiddleware]] = []
-
-        try:
-            for bot_cfg in tg_multi_cfg.bots:
-                auth = AuthMiddleware.from_bot_config(
-                    raw_config, "telegram", bot_cfg.bot_id, store=auth_store
-                )
-                if auth is None:
-                    log.warning(
-                        "telegram bot_id=%r has no auth config — skipping",
-                        bot_cfg.bot_id,
-                    )
-                    continue
-                tg_bot_auths.append((bot_cfg, auth))
-
-            for bot_cfg in dc_multi_cfg.bots:
-                auth = AuthMiddleware.from_bot_config(
-                    raw_config, "discord", bot_cfg.bot_id, store=auth_store
-                )
-                if auth is None:
-                    log.warning(
-                        "discord bot_id=%r has no auth config — skipping",
-                        bot_cfg.bot_id,
-                    )
-                    continue
-                dc_bot_auths.append((bot_cfg, auth))
-        except ValueError as exc:
-            sys.exit(str(exc))
-
-        if not tg_bot_auths and not dc_bot_auths:
-            sys.exit(
-                "No adapters configured — add at least one [[telegram.bots]] or"
-                " [[discord.bots]] entry with a matching [[auth.telegram_bots]] or"
-                " [[auth.discord_bots]] section to config.toml"
-            )
-
-        # Resolve (platform, bot_id) → agent_name: DB cache first, TOML fallback + seed.
-        bot_agent_map = await _resolve_bot_agent_map(
-            agent_store, tg_multi_cfg.bots, dc_multi_cfg.bots
-        )
-        agent_names: set[str] = set(bot_agent_map.values())
-
-        # Load all agent configs once (DB-first, TOML fallback for pre-migration
-        # compat). Used for cli_pool detection, smart routing, msg_manager, creation.
-        agent_configs: dict[str, Agent] = {}
-        for n in sorted(agent_names):
-            row = agent_store.get(n)
-            if row is not None:
-                agent_configs[n] = agent_row_to_config(
-                    row, instance_overrides=_build_agent_overrides(raw_config, n)
-                )
-            else:
-                # Agent not in DB — fall back to TOML file (pre-migration compatibility)
-                try:
-                    agent_configs[n] = load_agent_config(
-                        n, instance_overrides=_build_agent_overrides(raw_config, n)
-                    )
-                except Exception as exc:  # noqa: BLE001
-                    log.error("Failed to load agent %r: %s — skipping", n, exc)
-        if not agent_configs:
-            sys.exit(
-                "No agent configs could be loaded — run 'lyra agent init' to seed the"
-                " agents table, or ensure agents/*.toml files exist"
-            )
-        first_agent_name = next(iter(sorted(agent_configs)))
-        first_agent_config = agent_configs[first_agent_name]
-
-        # Use the first agent's language for messages (all agents share one msg_manager)
-        msg_manager = _load_messages(language=first_agent_config.i18n_language)
-
-        pairing_config = _load_pairing_config(raw_config)
-        if pairing_config.enabled and not admin_user_ids:
-            log.warning(
-                "Pairing is enabled but no admin user_ids configured in "
-                "[admin].user_ids — no one can run /invite or /unpair"
-            )
-        pm: PairingManager | None = None
-        if pairing_config.enabled:
-            pm = PairingManager(
-                config=pairing_config,
-                db_path=vault_dir_mb / "pairing.db",
-                admin_user_ids=admin_user_ids,
-                auth_store=auth_store,
-            )
-            await pm.connect()
-            set_pairing_manager(pm)
-
-        stt_service: STTService | None = None
-        if os.environ.get("STT_MODEL_SIZE"):
-            try:
-                # Env vars set baseline; agent TOML overlays on top
-                stt_cfg = apply_agent_stt_overlay(
-                    first_agent_config.stt, load_stt_config()
-                )
-                stt_service = STTService(stt_cfg)
-                log.info("STT enabled: model=%s (via voiceCLI)", stt_cfg.model_size)
-            except ValueError as exc:
-                raise SystemExit(f"Invalid STT configuration: {exc}") from exc
-
-        tts_service: TTSService | None = None
-        voice_responses = os.environ.get("LYRA_VOICE_RESPONSES", "1") != "0"
-        if stt_service is not None and voice_responses:
-            # Env vars set baseline; agent TOML overlays on top (first alphabetically)
-            tts_cfg = apply_agent_tts_overlay(first_agent_config.tts, load_tts_config())
-            # Warn if multiple agents define conflicting [tts] sections
-            _agents_with_tts = [
-                n for n, cfg in agent_configs.items() if cfg.tts is not None
-            ]
-            if len(_agents_with_tts) > 1:
-                log.warning(
-                    "Multiple agents define [tts] config: %s — "
-                    "using %r (first alphabetically). "
-                    "Other agents' [tts] settings are ignored.",
-                    _agents_with_tts,
-                    first_agent_name,
-                )
-            tts_service = TTSService(tts_cfg)
-            log.info(
-                "TTS voice responses enabled: engine=%s voice=%s",
-                tts_cfg.engine or "default",
-                tts_cfg.voice or "default",
-            )
-
-        from lyra.core.debouncer import DEFAULT_DEBOUNCE_MS
-
-        prefs_store = PrefsStore(db_path=vault_dir_mb / "auth.db")
-        await prefs_store.connect()
-
-        hub = Hub(
-            circuit_registry=circuit_registry,
-            msg_manager=msg_manager,
-            pairing_manager=pm,
-            stt=stt_service,
-            tts=tts_service,
-            debounce_ms=DEFAULT_DEBOUNCE_MS,
-            prefs_store=prefs_store,
-        )
-        hub.set_turn_store(turn_store_mb)
-
-        # Build cli_pool if any agent needs it (uses pre-loaded configs)
-        cli_pool: CliPool | None = None
-        for cfg in agent_configs.values():
-            if cfg.model_config.backend == "claude-cli":
-                cli_pool = CliPool()
-                await cli_pool.start()
-                break
-
-        # Create all unique agents — each gets its own ProviderRegistry built from
-        # its own smart_routing config inside _resolve_agents().
-        all_agents = _resolve_agents(
-            agent_configs,
-            cli_pool,
-            circuit_registry,
-            admin_user_ids,
-            msg_manager,
-            stt_service,
-            tts_service,
-        )
-        for ag in all_agents.values():
-            hub.register_agent(ag)
-
-        # Wire Telegram adapters
-        tg_adapters: list[TelegramAdapter] = []
-        tg_dispatchers: list[OutboundDispatcher] = []
-        for bot_cfg, auth in tg_bot_auths:
-            resolved_agent = bot_agent_map.get(("telegram", bot_cfg.bot_id))
-            if resolved_agent is None:
-                log.warning(
-                    "telegram bot_id=%r not in bot_agent_map — skipping adapter",
-                    bot_cfg.bot_id,
-                )
-                continue
-            tg_creds = await cred_store.get_full("telegram", bot_cfg.bot_id)
-            if tg_creds is None:
-                raise MissingCredentialsError("telegram", bot_cfg.bot_id)
-            tg_token, tg_webhook_secret = tg_creds
-            adapter = TelegramAdapter(
-                bot_id=bot_cfg.bot_id,
-                token=tg_token,
-                hub=hub,
-                bot_username=bot_cfg.bot_username,
-                webhook_secret=tg_webhook_secret or "",
-                circuit_registry=circuit_registry,
-                msg_manager=msg_manager,
-                auth=auth,
-            )
-            hub.register_adapter(Platform.TELEGRAM, bot_cfg.bot_id, adapter)
-            tg_key = RoutingKey(Platform.TELEGRAM, bot_cfg.bot_id, "*")
-            hub.register_binding(
-                Platform.TELEGRAM,
-                bot_cfg.bot_id,
-                "*",
-                resolved_agent,
-                tg_key.to_pool_id(),
-            )
-            dispatcher = OutboundDispatcher(
-                platform_name="telegram",
-                adapter=adapter,
-                circuit=circuit_registry.get("telegram"),
-                circuit_registry=circuit_registry,
-                bot_id=bot_cfg.bot_id,
-            )
-            hub.register_outbound_dispatcher(
-                Platform.TELEGRAM, bot_cfg.bot_id, dispatcher
-            )
-            tg_adapters.append(adapter)
-            tg_dispatchers.append(dispatcher)
-            log.info(
-                "Registered Telegram bot bot_id=%r agent=%r",
-                bot_cfg.bot_id,
-                resolved_agent,
-            )
-
-        # Wire Discord adapters
-        dc_adapters: list[tuple[DiscordAdapter, DiscordBotConfig, str]] = []
-        dc_dispatchers: list[OutboundDispatcher] = []
-        for bot_cfg, auth in dc_bot_auths:
-            resolved_agent = bot_agent_map.get(("discord", bot_cfg.bot_id))
-            if resolved_agent is None:
-                log.warning(
-                    "discord bot_id=%r not in bot_agent_map — skipping adapter",
-                    bot_cfg.bot_id,
-                )
-                continue
-            dc_creds = await cred_store.get_full("discord", bot_cfg.bot_id)
-            if dc_creds is None:
-                raise MissingCredentialsError("discord", bot_cfg.bot_id)
-            dc_token, _ = dc_creds
-            adapter = DiscordAdapter(
-                hub=hub,
-                bot_id=bot_cfg.bot_id,
-                circuit_registry=circuit_registry,
-                msg_manager=msg_manager,
-                auto_thread=bot_cfg.auto_thread,
-                auth=auth,
-                thread_store=thread_store_mb,
-            )
-            hub.register_adapter(Platform.DISCORD, bot_cfg.bot_id, adapter)
-            dc_key = RoutingKey(Platform.DISCORD, bot_cfg.bot_id, "*")
-            hub.register_binding(
-                Platform.DISCORD,
-                bot_cfg.bot_id,
-                "*",
-                resolved_agent,
-                dc_key.to_pool_id(),
-            )
-            dispatcher = OutboundDispatcher(
-                platform_name="discord",
-                adapter=adapter,
-                circuit=circuit_registry.get("discord"),
-                circuit_registry=circuit_registry,
-                bot_id=bot_cfg.bot_id,
-            )
-            hub.register_outbound_dispatcher(
-                Platform.DISCORD, bot_cfg.bot_id, dispatcher
-            )
-            dc_adapters.append((adapter, bot_cfg, dc_token))
-            dc_dispatchers.append(dispatcher)
-            log.info(
-                "Registered Discord bot bot_id=%r agent=%r",
-                bot_cfg.bot_id,
-                resolved_agent,
-            )
-
-        # Start buses and dispatchers
-        await hub.inbound_bus.start()
-        await hub.inbound_audio_bus.start()
-        for d in tg_dispatchers:
-            await d.start()
-        for d in dc_dispatchers:
-            await d.start()
-
-        health_port = int(os.environ.get("LYRA_HEALTH_PORT", "8443"))
-        health_app = create_health_app(hub)
-        health_config = uvicorn.Config(
-            health_app, host="127.0.0.1", port=health_port, log_level="warning"
-        )
-        health_server = uvicorn.Server(health_config)
-
-        stop = _stop if _stop is not None else asyncio.Event()
-        if _stop is None:
-            _loop = asyncio.get_running_loop()
-            _loop.add_signal_handler(signal.SIGINT, stop.set)
-            _loop.add_signal_handler(signal.SIGTERM, stop.set)
-
-        tasks = [
-            asyncio.create_task(hub.run(), name="hub"),
-            asyncio.create_task(hub._audio_loop(), name="hub-audio"),
-            asyncio.create_task(health_server.serve(), name="health"),
-        ]
-        for tg_adapter in tg_adapters:
-            tasks.append(
-                asyncio.create_task(
-                    tg_adapter.dp.start_polling(tg_adapter.bot, handle_signals=False),
-                    name=f"telegram:{tg_adapter._bot_id}",
-                )
-            )
-        for dc_adapter, dc_bot_cfg, dc_token in dc_adapters:
-            _dc_task = asyncio.create_task(
-                dc_adapter.start(dc_token),
-                name=f"discord:{dc_bot_cfg.bot_id}",
-            )
-            _dc_task.add_done_callback(_log_task_failure)
-            tasks.append(_dc_task)
-
-        tg_active = [f"telegram:{a._bot_id}" for a in tg_adapters]
-        dc_active = [f"discord:{c.bot_id}" for _, c, _ in dc_adapters]
-        active = tg_active + dc_active
-        log.info(
-            "Lyra started — adapters: %s, health on :%d.",
-            ", ".join(active) if active else "none",
-            health_port,
-        )
-
-        await stop.wait()
-
-        log.info("Shutdown signal received — stopping…")
-        for task in tasks:
-            task.cancel()
-        await asyncio.gather(*tasks, return_exceptions=True)
-        await hub.inbound_bus.stop()
-        await hub.inbound_audio_bus.stop()
-        for d in tg_dispatchers:
-            await d.stop()
-        for d in dc_dispatchers:
-            await d.stop()
-        for dc_adapter, _, _dc_tok in dc_adapters:
-            await dc_adapter.close()
-        if pm is not None:
-            await pm.close()
-        await cred_store.close()
-        await auth_store.close()
-        await agent_store.close()
-        await turn_store_mb.close()
-        if thread_store_mb is not None:
-            await thread_store_mb.close()
-        await prefs_store.close()
-        if cli_pool is not None:
-            await cli_pool.stop()
-        log.info("Lyra stopped.")
-    finally:
-        if cred_store is not None:
-            await cred_store.close()
-        if auth_store is not None:
-            await auth_store.close()
-        if agent_store is not None:
-            await agent_store.close()
-        if turn_store_mb is not None:
-            await turn_store_mb.close()
-        if thread_store_mb is not None:
-            await thread_store_mb.close()
-
-
-
-async def _bootstrap_legacy(  # noqa: C901, PLR0915 — startup wiring: each adapter/service requires sequential conditional setup
-    raw_config: dict,
-    circuit_registry: CircuitRegistry,
-    admin_user_ids: set[str],
-    *,
-    adapter: str = "all",
-    _stop: asyncio.Event | None = None,
-) -> None:
-    """Wire hub + adapters for the legacy single-bot schema and run until stop.
-
-    Handles [auth.telegram] / [auth.discord] flat sections from config.toml.
-    Behavior is identical to the original single-bot implementation.
-    """
-    # ------------------------------------------------------------------
-    # Legacy single-bot path (backward compat).
-    # Runs when no [[telegram.bots]] / [[discord.bots]] arrays are found.
-    # Behavior is identical to the original single-bot implementation.
-    # ------------------------------------------------------------------
-    vault_dir_lg = Path(os.environ.get("LYRA_VAULT_DIR", str(Path.home() / ".lyra")))
-    vault_dir_lg.mkdir(parents=True, exist_ok=True)
-    auth_store_legacy = AuthStore(db_path=vault_dir_lg / "auth.db")
-    await auth_store_legacy.connect()
-    keyring_lg = LyraKeyring.load_or_create(vault_dir_lg / "keyring.key")
-    cred_store_legacy = CredentialStore(
-        db_path=vault_dir_lg / "auth.db",
-        keyring=keyring_lg,
-    )
-    await cred_store_legacy.connect()
-    turn_store_lg = TurnStore(db_path=vault_dir_lg / "turns.db")
-    await turn_store_lg.connect()
-    thread_store_lg = ThreadStore(db_path=vault_dir_lg / "auth.db")
-    await thread_store_lg.connect()
-    await auth_store_legacy.seed_from_config(raw_config, "telegram")
-    await auth_store_legacy.seed_from_config(raw_config, "discord")
-
-    try:
-        tg_auth = (
-            None
-            if adapter == "discord"
-            else AuthMiddleware.from_config(
-                raw_config, "telegram", store=auth_store_legacy
-            )
-        )
-        dc_auth = (
-            None
-            if adapter == "telegram"
-            else AuthMiddleware.from_config(
-                raw_config, "discord", store=auth_store_legacy
-            )
-        )
-    except ValueError as exc:
-        sys.exit(str(exc))
-
-    if tg_auth is None and dc_auth is None:
-        sys.exit(
-            "No adapters configured — add at least one of [auth.telegram] or"
-            " [auth.discord] to config.toml"
-        )
-
-    # Resolve credentials from CredentialStore (bot_id="main" for legacy path).
-    # Non-credential config (bot_username, auto_thread) is read from env vars.
-    tg_token_lg: str | None = None
-    tg_webhook_secret_lg: str | None = None
-    tg_bot_username_lg: str = os.environ.get("TELEGRAM_BOT_USERNAME", "lyra_bot")
-    if tg_auth is not None:
-        tg_creds_lg = await cred_store_legacy.get_full("telegram", "main")
-        if tg_creds_lg is None:
-            raise MissingCredentialsError("telegram", "main")
-        tg_token_lg, tg_webhook_secret_lg = tg_creds_lg
-
-    dc_token_lg: str | None = None
-    dc_auto_thread_lg: bool = True
-    if dc_auth is not None:
-        dc_creds_lg = await cred_store_legacy.get_full("discord", "main")
-        if dc_creds_lg is None:
-            raise MissingCredentialsError("discord", "main")
-        dc_token_lg, _ = dc_creds_lg
-        auto_thread_str_lg = os.environ.get("DISCORD_AUTO_THREAD", "").strip().lower()
-        dc_auto_thread_lg = (
-            auto_thread_str_lg in {"1", "true", "yes", "on"}
-            if auto_thread_str_lg
-            else True
-        )
-
-    agent_config = load_agent_config(
-        "lyra_default",
-        instance_overrides=_build_agent_overrides(raw_config, "lyra_default"),
-    )
-    log.info(
-        "Agent loaded: name=%s model=%s backend=%s",
-        agent_config.name,
-        agent_config.model_config.model,
-        agent_config.model_config.backend,
-    )
-
-    msg_manager = _load_messages(language=agent_config.i18n_language)
-
-    pairing_config = _load_pairing_config(raw_config)
-    if pairing_config.enabled and not admin_user_ids:
-        log.warning(
-            "Pairing is enabled but no admin user_ids configured in "
-            "[admin].user_ids — no one can run /invite or /unpair"
-        )
-    pm_legacy: PairingManager | None = None
-    if pairing_config.enabled:
-        pm_legacy = PairingManager(
-            config=pairing_config,
-            db_path=vault_dir_lg / "pairing.db",
-            admin_user_ids=admin_user_ids,
-            auth_store=auth_store_legacy,
-        )
-        await pm_legacy.connect()
-        set_pairing_manager(pm_legacy)
-
-    stt_service_legacy: STTService | None = None
-    if os.environ.get("STT_MODEL_SIZE"):
-        try:
-            stt_cfg = load_stt_config()
-            stt_service_legacy = STTService(stt_cfg)
-            log.info("STT enabled: model=%s (via voiceCLI)", stt_cfg.model_size)
-        except ValueError as exc:
-            raise SystemExit(f"Invalid STT configuration: {exc}") from exc
-
-    tts_service_legacy: TTSService | None = None
-    _voice_responses_on = os.environ.get("LYRA_VOICE_RESPONSES", "1") != "0"
-    if stt_service_legacy is not None and _voice_responses_on:
-        tts_cfg = load_tts_config()
-        tts_service_legacy = TTSService(tts_cfg)
-        log.info(
-            "TTS voice responses enabled: engine=%s voice=%s",
-            tts_cfg.engine or "default",
-            tts_cfg.voice or "default",
-        )
-
-    from lyra.core.debouncer import DEFAULT_DEBOUNCE_MS
-
-    hub = Hub(
-        circuit_registry=circuit_registry,
-        msg_manager=msg_manager,
-        pairing_manager=pm_legacy,
-        stt=stt_service_legacy,
-        tts=tts_service_legacy,
-        debounce_ms=DEFAULT_DEBOUNCE_MS,
-    )
-    hub.set_turn_store(turn_store_lg)
-
-    cli_pool_legacy: CliPool | None = None
-    if agent_config.model_config.backend == "claude-cli":
-        cli_pool_legacy = CliPool()
-        await cli_pool_legacy.start()
-
-    sr_config = agent_config.smart_routing
-    if (
-        sr_config is not None
-        and sr_config.enabled
-        and agent_config.model_config.backend != "anthropic-sdk"
-    ):
-        log.warning(
-            "smart_routing.enabled=true but backend=%r — "
-            "smart routing only applies to anthropic-sdk",
-            agent_config.model_config.backend,
-        )
-    provider_registry, smart_routing_decorator = _build_provider_registry(
-        circuit_registry, cli_pool_legacy, smart_routing_config=sr_config
-    )
-
-    agent = _create_agent(
-        agent_config,
-        cli_pool_legacy,
-        circuit_registry=circuit_registry,
-        admin_user_ids=admin_user_ids,
-        msg_manager=msg_manager,
-        stt=stt_service_legacy,
-        tts=tts_service_legacy,
-        provider_registry=provider_registry,
-        smart_routing_decorator=smart_routing_decorator,
-    )
-    hub.register_agent(agent)
-
-    tg_adapter: TelegramAdapter | None = None
-    dc_adapter: DiscordAdapter | None = None
-
-    if tg_auth is not None and tg_token_lg is not None:
-        tg_adapter = TelegramAdapter(
-            bot_id="main",
-            token=tg_token_lg,
-            hub=hub,
-            bot_username=tg_bot_username_lg,
-            webhook_secret=tg_webhook_secret_lg or "",
-            circuit_registry=circuit_registry,
-            msg_manager=msg_manager,
-            auth=tg_auth,
-        )
-        hub.register_adapter(Platform.TELEGRAM, "main", tg_adapter)
-        tg_key = RoutingKey(Platform.TELEGRAM, "main", "*")
-        hub.register_binding(
-            Platform.TELEGRAM, "main", "*", agent.name, tg_key.to_pool_id()
-        )
-
-    if dc_auth is not None and dc_token_lg is not None:
-        dc_adapter = DiscordAdapter(
-            hub=hub,
-            bot_id="main",
-            circuit_registry=circuit_registry,
-            msg_manager=msg_manager,
-            auto_thread=dc_auto_thread_lg,
-            auth=dc_auth,
-            thread_store=thread_store_lg,
-        )
-        hub.register_adapter(Platform.DISCORD, "main", dc_adapter)
-        dc_key = RoutingKey(Platform.DISCORD, "main", "*")
-        hub.register_binding(
-            Platform.DISCORD, "main", "*", agent.name, dc_key.to_pool_id()
-        )
-
-    # Create and register per-platform outbound dispatchers
-    tg_dispatcher_legacy: OutboundDispatcher | None = None
-    dc_dispatcher_legacy: OutboundDispatcher | None = None
-
-    if tg_adapter is not None:
-        tg_dispatcher_legacy = OutboundDispatcher(
-            platform_name="telegram",
-            adapter=tg_adapter,
-            circuit=circuit_registry.get("telegram"),
-            circuit_registry=circuit_registry,
-            bot_id="main",
-        )
-        hub.register_outbound_dispatcher(
-            Platform.TELEGRAM, "main", tg_dispatcher_legacy
-        )
-
-    if dc_adapter is not None:
-        dc_dispatcher_legacy = OutboundDispatcher(
-            platform_name="discord",
-            adapter=dc_adapter,
-            circuit=circuit_registry.get("discord"),
-            circuit_registry=circuit_registry,
-            bot_id="main",
-        )
-        hub.register_outbound_dispatcher(Platform.DISCORD, "main", dc_dispatcher_legacy)
-
-    # Start InboundBus feeder tasks and OutboundDispatcher workers
-    await hub.inbound_bus.start()
-    await hub.inbound_audio_bus.start()
-    if tg_dispatcher_legacy is not None:
-        await tg_dispatcher_legacy.start()
-    if dc_dispatcher_legacy is not None:
-        await dc_dispatcher_legacy.start()
-
-    # Health endpoint (SC-1): root FastAPI app on configurable port
-    health_port = int(os.environ.get("LYRA_HEALTH_PORT", "8443"))
-    health_app = create_health_app(hub)
-    health_config = uvicorn.Config(
-        health_app, host="127.0.0.1", port=health_port, log_level="warning"
-    )
-    health_server = uvicorn.Server(health_config)
-
-    stop = _stop if _stop is not None else asyncio.Event()
-    if _stop is None:
-        _loop = asyncio.get_running_loop()
-        _loop.add_signal_handler(signal.SIGINT, stop.set)
-        _loop.add_signal_handler(signal.SIGTERM, stop.set)
-
-    tasks = [
-        asyncio.create_task(hub.run(), name="hub"),
-        asyncio.create_task(hub._audio_loop(), name="hub-audio"),
-        asyncio.create_task(health_server.serve(), name="health"),
-    ]
-    if tg_adapter is not None:
-        tasks.append(
-            asyncio.create_task(
-                # handle_signals=False is mandatory when aiogram runs inside
-                # asyncio.gather() — otherwise it installs its own signal handlers
-                # that conflict with the event loop's, causing double-cancellation.
-                tg_adapter.dp.start_polling(tg_adapter.bot, handle_signals=False),
-                name="telegram",
-            )
-        )
-    if dc_adapter is not None and dc_token_lg is not None:
-        _dc_task_lg = asyncio.create_task(dc_adapter.start(dc_token_lg), name="discord")
-        _dc_task_lg.add_done_callback(_log_task_failure)
-        tasks.append(_dc_task_lg)
-
-    task_names = {t.get_name() for t in tasks}
-    active = [name for name in ("telegram", "discord") if name in task_names]
-    log.info(
-        "Lyra started — adapters: %s, health on :%d.",
-        ", ".join(active) if active else "none",
-        health_port,
-    )
-
-    await stop.wait()
-
-    log.info("Shutdown signal received — stopping…")
-    for task in tasks:
-        task.cancel()
-    await asyncio.gather(*tasks, return_exceptions=True)
-    await hub.inbound_bus.stop()
-    await hub.inbound_audio_bus.stop()
-    if tg_dispatcher_legacy is not None:
-        await tg_dispatcher_legacy.stop()
-    if dc_dispatcher_legacy is not None:
-        await dc_dispatcher_legacy.stop()
-    if dc_adapter is not None:
-        await dc_adapter.close()
-    if pm_legacy is not None:
-        await pm_legacy.close()
-    await cred_store_legacy.close()
-    await auth_store_legacy.close()
-    await turn_store_lg.close()
-    await thread_store_lg.close()
-    if cli_pool_legacy is not None:
-        await cli_pool_legacy.stop()
-    log.info("Lyra stopped.")
 
 
 def _parse_args(args: list[str] | None = None) -> argparse.Namespace:
