+++ src/lyra/bootstrap/__init__.py
@@ -1,0 +1,1 @@
+"""Bootstrap package — startup wiring helpers extracted from __main__."""
